fix relative import in tests from a previous refactor
