+++ tests/test_op_api/group/list/test_group_list_all.py
@@ -12,7 +12,7 @@ from pyonepassword.api.exceptions import OPInvalidGroupListException
 # this also reduced exercising tested code simply by importing
 if TYPE_CHECKING:
     from pyonepassword import OP
-    from ..fixtures.expected_group_data import (
+    from ....fixtures.expected_group_data import (
         ExpectedGroupListData,
         ExpectedGroupListEntry
     )
