+++ tests/test_op_api/group/list/test_group_list_vault.py
@@ -9,7 +9,7 @@ import pytest
 # this also reduced exercising tested code simply by importing
 if TYPE_CHECKING:
     from pyonepassword import OP
-    from ..fixtures.expected_group_data import (
+    from ....fixtures.expected_group_data import (
         ExpectedGroupListData,
         ExpectedGroupListEntry
     )
