+++ tests/test_op_api/item/edit/test_item_edit.py
@@ -10,7 +10,10 @@ if TYPE_CHECKING:
     from pyonepassword import OP
     from pyonepassword.api.object_types import OPLoginItem
 
-    from ..fixtures.expected_login import ExpectedLogin, ExpectedLoginItemData
+    from ....fixtures.expected_login import (
+        ExpectedLogin,
+        ExpectedLoginItemData
+    )
 
 # ensure HOME env variable is set, and there's a valid op config present
 pytestmark = pytest.mark.usefixtures("valid_op_cli_config_homedir")
