+++ tests/test_op_api/item/list/test_item_list_ssh_key.py
@@ -5,7 +5,7 @@ from typing import TYPE_CHECKING
 import pytest
 
 if TYPE_CHECKING:
-    from ..fixtures.expected_ssh_key_data import ExpectedSSHKey, ExpectedSSHKeyData
+    from ....fixtures.expected_ssh_key_data import ExpectedSSHKey, ExpectedSSHKeyData
     from pyonepassword import OP
 
 from pyonepassword.api.descriptor_types import OPSSHKeyItemDescriptor
